+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 7.0.0
 
 - **BREAKING** Change function signature to get rid of `promptUser`
-- **BREAKING** Change format of `package-deps` in manifest from `"my-name:my-version"` to be `{ "name": "my-name", "version": "my-version" }`
+- **BREAKING** Change format of `package-deps` in manifest from `"my-name:my-version"` to be `{ "name": "my-name", "minimumVersion": "my-version" }`
+- **BREAKING** Instead of accepting a semver-range for version, you can now only specify a minimum version in `x.y.z` format, instead of for example `>=2`
 - Add support for optional/group packages where users only have to install one of many, and are given a choice at runtime. Signature becomes `Dependency[][]` instead of `Dependency[]`
 - Add support for using through CLI for CI and similar purposes